feat(core): Add a start method to run nutra
Running nutra directly from its constructor doesn't
seem to be an issue now, but it will cause issues
in the future when the user has ability to interact
with the public api (i.e. listen to events). Also
helps in unit testing the core since the constructor
can only be initiated once, thus prohibiting the use
of spies.