refactor(core): Don't allow empty strings into option resolution
This is referring to the private "getRequiredOptions" method
and this change favors returning empty strings, if one was
provided. Originally, it would try to resolve the empty string
path and require it. This fails and the function returns null,
which isn't a bit deal, but it makes more sense to reject
empty inputs upfront.