refactor(core): Make plugin hooks inextendable
This is in referrence to the private "getPluginHooks" method.
It is responsible for fishing hooks from the plugin constructors.
Right now it just passes in a default object, and it appears to
be a good idea to pass in an inextendable object. Mainly to avoid
issues with users trying to set an event that doesn't exist, or
modify existing ones.