fix(core): Make temporary directory persist throughout unit tests
One of the unit tests makes sure that nutra deletes the temporary
directory on system exit, however this is the same temporary
directory that might also be in use by the master nutra instance.
So instead of deleting the temporary folder completely, the
system now has a constant temporary folder and with sub
directories for each instance.