refactor: Improve debugging experience
This commit improves the debugging experience by ensuring that
fatal warnings show the stack, that configuration
resolution throws a machine generated error if it does not pertain
to the config file not being found, and that invalid globs error
messages are more detailed. On the dev side of things, paths are
now normalized in less places and some spelling mistakes were
fixed — making the dev experience a lil better.