fix: Ensure "nutra-jspm" is compatible with nutra
The "nutra-coverage" module loader needs access to a physical
transpiled file in order for to do its work, which then dawned on
me; who's reponsibility is it to save the file to disk? Well, it
appears the best answer atm is; "whoever needs them". So cache is
now compartmentalized by plugin names, which write to their own
folder.