feat: Ability to provide absolute paths
Previously, all paths (i.e. config file, preprocessor, and file
patterns) had to be relative to the current working directory. This
fits most use cases, but more complex one with custom enviroments
need a bit more flexibility. This commit does not provide the
ability to pass in an absolute config url to the CLI.