refactor: Add mechanism to detect uncaught/unhandled errors
Not matter how airtight this seems to be, there is always a silent
error creeping up here in there. Node "domains" would have been
a decent way to ensure all errors are handled, but it's being
deprecated. The only way of truly detecting uncaught exceptions and
unhandled rejections seems to be by listening to the events on a
global level. A module called "gotta-catch-em-all" was introduced
to do exactly that.