fix: Ensure CLI exits with the correct code
Previously, it gave that reponsibility to the core, which in theory
worked as long as the core exited an explicity error, however this
is not always the case. This commits still lets the core deal with
the exit if possible, but the CLI will exit with the code provided
by the core if it fails to exit.